[CodeCompletion] Move solver-based key path completion to its own file
