+++ include/swift/IDE/KeyPathCompletion.h
@@ -1,0 +1,54 @@
+//===--- KeyPathCompletion.h ----------------------------------------------===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2014 - 2022 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+// See https://swift.org/CONTRIBUTORS.txt for the list of Swift project authors
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef SWIFT_IDE_KEYPATHCOMPLETION_H
+#define SWIFT_IDE_KEYPATHCOMPLETION_H
+
+#include "swift/IDE/CodeCompletionConsumer.h"
+#include "swift/IDE/CodeCompletionContext.h"
+#include "swift/Sema/CodeCompletionTypeChecking.h"
+
+namespace swift {
+namespace ide {
+
+class KeyPathTypeCheckCompletionCallback : public TypeCheckCompletionCallback {
+public:
+  struct Result {
+    /// The type on which completion should occur, i.e. a result type of the
+    /// previous component.
+    Type BaseType;
+    /// Whether code completion happens on the key path's root.
+    bool OnRoot;
+  };
+
+private:
+  KeyPathExpr *KeyPath;
+  SmallVector<Result, 4> Results;
+
+public:
+  KeyPathTypeCheckCompletionCallback(KeyPathExpr *KeyPath) : KeyPath(KeyPath) {}
+
+  ArrayRef<Result> getResults() const { return Results; }
+
+  void sawSolution(const constraints::Solution &solution) override;
+};
+
+void deliverKeyPathResults(
+    ArrayRef<KeyPathTypeCheckCompletionCallback::Result> Results,
+    DeclContext *DC, SourceLoc DotLoc,
+    ide::CodeCompletionContext &CompletionCtx,
+    CodeCompletionConsumer &Consumer);
+
+} // end namespace ide
+} // end namespace swift
+
+#endif // SWIFT_IDE_KEYPATHCOMPLETION_H
