+++ lib/Sema/TypeCheckCodeCompletion.cpp
@@ -850,57 +850,3 @@ bool swift::isImplicitSingleExpressionReturn(ConstraintSystem &CS,
   }
   return false;
 }
-
-void KeyPathTypeCheckCompletionCallback::sawSolution(
-    const constraints::Solution &S) {
-  // Determine the code completion.
-  size_t ComponentIndex = 0;
-  for (auto &Component : KeyPath->getComponents()) {
-    if (Component.getKind() == KeyPathExpr::Component::Kind::CodeCompletion) {
-      break;
-    } else {
-      ComponentIndex++;
-    }
-  }
-  assert(ComponentIndex < KeyPath->getComponents().size() &&
-         "Didn't find a code compleiton component?");
-
-  Type BaseType;
-  if (ComponentIndex == 0) {
-    // We are completing on the root and need to extract the key path's root
-    // type.
-    if (KeyPath->getRootType()) {
-      BaseType = S.getResolvedType(KeyPath->getRootType());
-    } else {
-      // The key path doesn't have a root TypeRepr set, so we can't look the key
-      // path's root up through it. Build a constraint locator and look the
-      // root type up through it.
-      // FIXME: Improve the linear search over S.typeBindings when it's possible
-      // to look up type variables by their locators.
-      auto RootLocator =
-          S.getConstraintLocator(KeyPath, {ConstraintLocator::KeyPathRoot});
-      auto BaseVariableTypeBinding =
-          llvm::find_if(S.typeBindings, [&RootLocator](const auto &Entry) {
-            return Entry.first->getImpl().getLocator() == RootLocator;
-          });
-      if (BaseVariableTypeBinding != S.typeBindings.end()) {
-        BaseType = S.simplifyType(BaseVariableTypeBinding->getSecond());
-      }
-    }
-  } else {
-    // We are completing after a component. Get the previous component's result
-    // type.
-    BaseType = S.simplifyType(S.getType(KeyPath, ComponentIndex - 1));
-  }
-  if (BaseType.isNull()) {
-    return;
-  }
-
-  // If ExpectedTy is a duplicate of any other result, ignore this solution.
-  if (llvm::any_of(Results, [&](const Result &R) {
-    return R.BaseType->isEqual(BaseType);
-  })) {
-    return;
-  }
-  Results.push_back({BaseType, /*OnRoot=*/(ComponentIndex == 0)});
-}
